// frontend/Angular/src/app/Models/twitter-user-profile.model.ts
export interface TwitterUserProfile {
  screenName: string;
  UserName: string;
  Location: string;
  Description: string;
  Email: string;
  profileImageUrl: string;
  followersCount: string;
  tweetCount: string;
  followingCount: string;
  mediaCount: string;
  likeCount:string; 
  TwitterUID:string;
}